requirements.txt
streamlit==1.32.2
plotly==6.0.0
pandas==2.2.1
numpy==1.26.4
folium==0.16.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
haversine==2.9.0
streamlit-folium==0.24.0
Pillow==10.2.0